keys_impl: add "Keyring" interface
Add a" Keyring: interface with the idea to use one common
"load_private_keys()" implementations that could be called
from both "InsecureFileKeyring" and "EnvVarKeyring" objects without
the need to repeat the same code twice.

Additionally, if tomorrow a new way of loading the keys is created
it will only need to inherit "Keyring" and implement an internal
function "_load_key()".

Signed-off-by: Martin Vrachev <mvrachev@vmware.com>