test_cli: test changes in targets delegations
Add testing about setting different incompatible scenarios like using
standard delegated roles and succinct hash bin delegation.
The idea is to make sure there is no situation in which through
the commands one can create a non-verifiable targets file.

The only states that are currently allowed can produce a
non-verifiable targets file are related to the decision to split
"add-succinct-delegation" and "init-succinct-roles" commands in two
instead of using one command.
This was done on purpose so that the CLI users can understand better
what is exactly happening.

Signed-off-by: Martin Vrachev <mvrachev@vmware.com>