Tests: fix expected error handling
Currently, if you want to test an expected error message as a result
from subprocess execution, we are comparing if the given error message
is identical to the one received.
That's not good as error messages can be long and contain locally bound
specific information.

It's better if we check that a specific error message is raised instead
of doing one-to-one comparison.

Additionally, the process return code should be handled differently
when there is an error and when the processes have been completed
successfully.

Signed-off-by: Martin Vrachev <mvrachev@vmware.com>