Re-design target addition
Adding a target only modifies a single metadata so being a "edit" sub
command makes sense. However, often user does not know which metadata
should be modified: this is especially the case with succinct
delegations.

Make add-target a top level action instead of a edit subaction:
* role can still be optionally specified but it is a starting point
  for the search
* delegation tree is followed until the first leaf role (that does
  not delegate further) is found
* command now outputs which role it inserted the target file into

This means add-target is now a simple repository wide command which
seems quite logical: user does not need to know the details, they just
want to add a target file.