Refactor aborting edits
The edit() context manager is powerful but canceling an edit
from inside the context manager with-block was clunky.

Make sure the edit() implementation handles AbortEdit exception:
This way the caller can raise AbortEdit if they want to cancel
whatever they're doing, without having to wrap everything inside a
try-catch block. It's not very discoverable but works quite well.

Example:

    with repo.edit("snapshot") as snapshot:
        # ... figure out here if snapshot actually needs changes

        if not changes_needed:
            raise AbortEdit("skipping unneeded snapshot")

    print("Snapshot process done: maybe saved a new version")

Here snapshot may get a new version or not, and code after the
with-block executes normally in either case.

Signed-off-by: Jussi Kukkonen <jkukkonen@vmware.com>