Refactor target adding
librepos Repository is the generic implementation that does not know
about storage. GitRepository is a Repository that uses git as the
storage method.

Refactor the 'add-target' and 'remove-target' implementations so that
this division between Repository and GitRepository is respected:
 * remove_target is 100% implemented in Repository
   (because remove_target does not, at least currently, handle removing
   any actual target files even if the git repository contains them)
 * add_target is implemented in Repository but GitRepository has a
   small override to handle adding actual target files and
   hash-prefixed symlinks into git

remove_target was moved as is, add_target had a minor simplification
done as it was split.