cli.py: split key creation from succinct init
The current succinct delegation process has three parts:

1. create a delegation
2. create a delegation key (store it for each delegated role)
3. initialize metadata for each delegated role

but only two commands:
1. tufrepo edit demo add-delegation --succinct 256 demo-bin
2. tufrepo init-succinct-roles demo

This works, but the key generation happens internally and is magic for
our users. This is a problem as the user doesn't have control over
the key generation process (for example how many keys does he want to
use for succinct hash bin delegation) and it's not a good practice given
how important is the key generation step.

As a solution, remove the key generation logic from the second command
"init-succinct-roles" and instead move it to "add-key".

Now the full succinct process can be achieved with three commands:
1. tufrepo edit demo add-delegation --succinct 256 demo-bin
2. tufrepo edit demo add-key
3. tufrepo init-succinct-roles demo

Signed-off-by: Martin Vrachev <mvrachev@vmware.com>