cli: Separate the private features better
Storing the private key should be a separate things from
storing the uri (the details of how the key is accessed):
This will make it easier to add Yubikey support.

Now --online will embed the URI in the metadata (by default URI goes to
a local file).

Test changes because now it is add_key() that fails when trying to add a
key without delegate in a case where succinct delegations are not used.